fix integer types for port numbers/nanoseconds
Motivation:

In at least two places we picked the integer types in an unfortunate
way:

- `TimeAmount` stores its nano-seconds amount in an `Int` which is fine
   for 64-bit platforms but not for 32-bit (2^32 ns is just over 4 s).
   Fortunately we could define `TimeAmount.Value` which is `Int` on 64-bit
   platforms (to not break the API) but `Int64` on 32-bit platforms.
   Ideally it would just be `Int64` everywhere
- We always use `Int` for port numbers, except the `port` value of
  `SocketAddress` uses `UInt16` for no good reason

Modifications:

- changed `SocketAddress#port` to be `Int`
- changed `TimeAmount` to always work with `Int64`

Result:

- more consistency

+++ Tests/NIOTests/EmbeddedEventLoopTest.swift
@@ -110,7 +110,7 @@ public class EmbeddedEventLoopTest: XCTestCase {
         var sentinel = 0
         let loop = EmbeddedEventLoop()
         for index in 1...10 {
-            _ = loop.scheduleTask(in: .nanoseconds(index)) {
+            _ = loop.scheduleTask(in: .nanoseconds(Int64(index))) {
                 sentinel = index
             }
         }
